Flexible window attributes
User can now specify the window attributes before the creation of the window itself. To enhance the API even more, I'm planning to make the function for updating the attribute list and the configuration of the window itself, but for now I wanted to get the basics of the way.